fuel-map/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
aiohttp==3.9.3
python-dotenv==1.0.1
